refactor(home): extract component logic to a custom hook
